+++ examples/graph_example/main.rs
@@ -24,10 +24,10 @@ async fn main() {
     // Connect to database and generates collections and indexes
     let db_connection = DatabaseConnection::builder()
         .with_credentials(
-            &std::env::var("DB_HOST").unwrap_or(DEFAULT_DB_HOST.to_string()),
-            &std::env::var("DB_NAME").unwrap_or(DEFAULT_DB_NAME.to_string()),
-            &std::env::var("DB_USER").unwrap_or(DEFAULT_DB_USER.to_string()),
-            &std::env::var("DB_PWD").unwrap_or(DEFAULT_DB_PWD.to_string()),
+            &std::env::var("DB_HOST").unwrap_or_else(|_| DEFAULT_DB_HOST.to_string()),
+            &std::env::var("DB_NAME").unwrap_or_else(|_| DEFAULT_DB_NAME.to_string()),
+            &std::env::var("DB_USER").unwrap_or_else(|_| DEFAULT_DB_USER.to_string()),
+            &std::env::var("DB_PWD").unwrap_or_else(|_| DEFAULT_DB_PWD.to_string()),
         )
         .with_auth_mode(AuthMode::Jwt)
         .with_schema_path("examples/graph_example/schema.yaml")
@@ -222,7 +222,6 @@ async fn main() {
             .unwrap();
     assert_eq!(
         children
-            .documents
             .iter()
             .map(|r| r.id().as_str())
             .collect::<Vec<&str>>(),
@@ -235,7 +234,6 @@ async fn main() {
             .unwrap();
     assert_eq!(
         children
-            .documents
             .iter()
             .map(|r| r.id().as_str())
             .collect::<Vec<&str>>(),
@@ -251,7 +249,6 @@ async fn main() {
     .unwrap();
     assert_eq!(
         ancestors
-            .documents
             .iter()
             .map(|r| r.id().as_str())
             .collect::<Vec<&str>>(),
@@ -272,7 +269,6 @@ async fn main() {
     .unwrap();
     assert_eq!(
         relatives
-            .documents
             .iter()
             .map(|r| r.id().as_str())
             .collect::<Vec<&str>>(),